Proxy requests
This adds two packages: proxy and server.

The server package runs the HTTP server that listens to all requests.
At the moment it just forwards requests to the proxy server.

The proxy package contains utility functions for request proxying.
At the moment it handles requests and forwards them on to a given URL.

A request goes through main -> server -> proxy.

I've also added doc/architecture_decisions to help me keep track of
what I intend to do.

The next step will be to add some configuration controls. It would be
nice for a user of the tool to be able to specify the URL of the
service they'd like requests to be forwarded to, for instance.