Request caching
This adds the bones of a caching feature, enabling requests
to be cached in an external cache.

The implementation is a trade off: Unlike Varnish, Shiny has
an external cache service (memcached only at the moment).
This gives flexibility (allows Shiny to talk to various caches)
and simplicity, as it can focus on being a web accelerator.
It sacrifices some speed for this, as it needs to talk to another
service, and reliability, as two things need to be running.
It might be nice for there to be an internal cache at some point,
to solve that.

There is further work needed on this feature. It should only
cache some requests (e.g. GET/HEAD requests with various codes,
I don't need to have too much customization over this right now).

It should also set expirations and so on.